docs: add a deprecation notice for node v8 to the code and readme
